+++ lib/src/api/list_fonts.dart
@@ -6,7 +6,7 @@ import 'constants.dart';
 
 /// Returns a list of font ids.
 Future<List<String>> listFonts() async {
-  var response = await http.get(Uri.parse('$apiUrl/fontlist'));
+  final response = await http.get(Uri.parse('$apiUrl/fontlist'));
   Map<String, String> fontList =
       jsonDecode(response.body).cast<String, String>();