+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.0
+
+- Move font configuration into `fonts` key in fontsource config.
+- Move generated package directory to `.fontsource`.
+- Fix current working directory resolving incorrectly.
+
 ## 0.1.1
 
 - Fix version key not accepting "latest".